+++ common/yak/ssa/program.go
@@ -139,6 +139,36 @@ func (prog *Program) IsPackagePathInList(pkgName string) bool {
 	return false
 }
 
+func (p *Program) GetEditor(url string) (*memedit.MemEditor, bool) {
+	return p.editorMap.Get(url)
+}
+func (p *Program) PushEditor(e *memedit.MemEditor) {
+	p.editorStack.Push(e)
+	p.editorMap.Set(e.GetUrl(), e)
+}
+
+func (p *Program) GetIncludeFiles() []string {
+	return p.editorMap.Keys()
+}
+
+func (p *Program) getCurrentEditor() *memedit.MemEditor {
+	if p.editorStack == nil || p.editorStack.Len() <= 0 {
+		return nil
+	}
+	_, v, ok := p.editorStack.Last()
+	if !ok {
+		return nil
+	}
+	return v
+}
+
+func (p *Program) PopEditor() {
+	if p.editorStack == nil || p.editorStack.Len() <= 0 {
+		return
+	}
+	p.editorStack.Pop()
+}
+
 func NewPackage(name string) *Package {
 	pkg := &Package{
 		Name:  name,
